requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.1
imageio==2.37.0
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.5
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyfingerprint==1.5
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
scikit-image==0.25.2
scipy==1.16.1
six==1.17.0
tifffile==2025.6.11
